services/backend/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==4.9.0
asyncpg==0.27.0
bcrypt==4.0.1
click==8.1.8
dictdiffer==0.9.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.88.0
h11==0.14.0
idna==3.10
iso8601==1.1.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==1.10.21
pypika-tortoise==0.1.6
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2025.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.22.0
tomlkit==0.13.2
tortoise-orm==0.19.2
typing_extensions==4.13.1
uvicorn==0.20.0